% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/configDropConnect.R
\name{configDropConnect}
\alias{configDropConnect}
\title{Configurate your connection between your app and DropBox}
\usage{
configDropConnect(token)
}
\arguments{
\item{token}{the path to get to your token (it must be accesible for your app)}
}
\value{
nothing
}
\description{
Configurate your connection between your app and DropBox
}
