% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/firstConnect.R
\name{firstConnect}
\alias{firstConnect}
\title{Authorize your application to get to your dropbox}
\usage{
firstConnect(dbFolder, tokenFile)
}
\arguments{
\item{dbFolder}{The folder that will contain your files}

\item{tokenFile}{The path where you want to stock your token in your app}
}
\value{
nothing
}
\description{
Authorize your application to get to your dropbox
}
