% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getToken.R
\name{getToken}
\alias{getToken}
\title{Title}
\usage{
getToken(tokenFile)
}
\arguments{
\item{tokenFile}{The path where you stock your token in your app}
}
\value{
your token that must stock in you app
}
\description{
Title
}
