% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadDrop.R
\name{loadDrop}
\alias{loadDrop}
\title{Load the data that you want visualize in your app}
\usage{
loadDrop(data, dbFolder)
}
\arguments{
\item{data}{the name of the data that you want bring in your app}

\item{dbFolder}{the directory name containing your data file in dropbox}
}
\value{
your data
}
\description{
Load the data that you want visualize in your app
}
