% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saveDrop.R
\name{saveDrop}
\alias{saveDrop}
\title{Save your work in dropBox}
\usage{
saveDrop(newData, dataStore, nameFile, dbFolder)
}
\arguments{
\item{newData}{the data containing the new information that you want stock}

\item{dataStore}{the Object (dataFrame or matrix) containing your data in R}

\item{nameFile}{the name of the file that contain you data in dropbox}

\item{dbFolder}{the directory name containing your data file in dropbox}
}
\value{
nothing
}
\description{
Save your work in dropBox
}
